requirements.txt
flask==3.0.0
flask-login==0.6.3
pandas==2.1.4
sqlalchemy==2.0.23
pymysql==1.1.0
reportlab==4.0.7
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.0.1
